pyproject.toml
[tool.poetry]
name = "repl-nix-historyapp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_historyapp"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
google-generativeai = "^0.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
